Moved dxf translator to ../lib/dxf.js
